Support Path-like objects
This change allows passing Path-like objects to most functions in blobfile.
- Local path-like objects are supported using the __fspath__ protocol from PEP 519.
  - This adds support for pathlib.Path.
- Remote path-like objects are supported using the __blobpath__ protocol proposed by @hauntsaninja.
  - Using a separate protocol for these means that a RemotePath class can define it and be passed to blobfile without type checkers thinking that it can also be passed to os.path.exists etc., as would be the case if it implemented __fspath__.

Objects implementing these protocols are immediately converted to strings by invoking the appropriate method as soon as they are passed to blobfile. Strings are passed through unchanged.

We use Protocol to define the typing and identify objects supporting __blobpath__; this requires Python 3.8. The implementation can be changed to support older versions of Python if desired.