Adds optional azure-identity support (#210)
`blobfile` currently uses a custom implementation of auth to Azure
Storage. It works with a client secret, but doesn't work in other
(recommended) credential-free flows, such as

* Managed Identity - apps talk to the local IMDS/IDENTITY_ENDPOINT for a
token associated with a VM/Web App's identity
* Workload Identity - apps exchange an OIDC token (ex: minted by an AKS
cluster) for an AAD token associated with a particular User Assigned
Identity

This PR adds an opt-in feature with no new dependencies. If apps set
`AZURE_USE_IDENTITY=1`, `blobfile` will bypass the custom implementation
here and instead use
[`DefaultAzureCredential`](https://learn.microsoft.com/en-us/python/api/azure-identity/azure.identity.defaultazurecredential?view=azure-python)
to retrieve tokens - which has support for the above named flows (and
more). Apps will need to bring their own `azure-identity` dependency via
`pyproject.toml`, `requirements.txt`, or otherwise

This PR is a backport of the same feature which was recently added in
`boostedblob`: https://github.com/hauntsaninja/boostedblob/pull/19