Improve exception handling
This improves exception chaining and fixes a case where we'd get a
BufferError after an exception. The bad case is if you have a finally
block that writes to the file after an exception, e.g. if you use a gzip
file in a context manager it will write a trailer to the underlying file
on close.

Here is a test (will commit once blobfile tests work again):
```
import blobfile
import contextlib
import unittest.mock

@contextlib.contextmanager
def trailer_file(fn: str):
    bf = blobfile.create_context(azure_write_chunk_size=1024)
    with bf.BlobFile(fn, "wb", streaming=True) as f:
        try:
            yield f
        finally:
            f.write(b"asdf" * 1024)

def main():
    with trailer_file("az://oaishantanu/mess/xx.tmp") as f:
        with unittest.mock.patch("blobfile._azure.execute_api_request") as mock:
            mock.side_effect = RuntimeError("asdf")
            f.write(b"asdf")
            f.write(b"asdf" * 1024 * 2)
            assert mock.call_count == 1

main()
```