Improve exception handling (#230)
This improves exception chaining and fixes a case where we'd get a
BufferError after an exception. The bad case is if you have a finally
block that writes to the file after an exception, e.g. if you use a gzip
file in a context manager it will write a trailer to the underlying file
on close.