Fix glob for blobs with unicode characters
Azure promises the following:

> Blobs are listed in alphabetical order in the response body, with upper-case letters listed first.

Of course it turns out that Azure has a weird definition of
alphabetical. Based on writing a blob with every Unicode character,
it appears to be ordered by UTF-16-BE

I'm not sure what GCS does, so I opted to just remove the assert. In
general, implicit dirs are sort of problematic and not too useful.