Add option to support blind writes (#254)
Adds `use_blind_writes` and the corresponding env var
`BLOBFILE_USE_BLIND_WRITES`. If the env var is set (and not overridden
by the config) or if the config value is set to True, this causes
blobfile to skip logic in the write path for Azure blobs that would
require read permissions:

- skips the check that the discovered credentials actually grant read
access to the container
- skips the check for too many uncommitted blocks / existing blob of
wrong type

Tested locally against an Azure container to confirm that with this set
you can write blobs without read perms.