Avoid GetBlobProperties calls when reading from Azure
When doing reads from Azure, avoid an extra call to get the size of the
file before we start reading it. Instead, just start reading in chunks;
if we try to read past the end of the file, we'll get a partial chunk,
and if there's no content we'll get a 416 error, both of which are
already handled gracefully.

Given that this change changes the user-visible behavior, in that
FileNotFoundError now isn't raised until we actually try to read from
the file, add a new option to the configuration, `eagerly_get_size`,
which defaults to `False` but can be set to `True` to revert to the old
behavior via the `configure()` function or by setting the env var
`BLOBFILE_EAGERLY_GET_SIZE` to `1`.