Add abort_on_exception option for BlobFile
Without using this feature, if one writes code like
```
  try:
    with blobfile.BlobFile(path, "w", streaming=True) as f:
      f.write("Hello, world!")
      raise Exception("oops!")
  except Exception as e:
```

Then there is the (possibly surprising) behavior of writing out an
empty blob to `path`. This change gives the option of adding
`abort_on_exception=True` to the BlobFile constructor to avoid this
behavior.

I considered instead making the option require the user to explicitly
call `close()` on the object, but that seemed less clean to me. I'm
open to other ideas/suggestions though.

Note that for `file://` blobs, we specifically reject the use of
`abort_on_exception=True`, because all other options seem too
scary. We could do one of:

 1. completely ignore the option
 2. Write to a different named blob and then do rename() for commit
 3. Buffer all writes in memory and write them all out at commit

And each one of those options feels/seems too scary to me - thus the
choice of throwing an error.